Day 7 part b
OK that was a pain in the neck, kinda. Unlike the previous ones which needed a
slight tweak, this needed a full on DFS implementation as well as detection
for the odd one out. Since we were given the clue that only one entry is wrong
then either all of the entries match, or one entry does not match. That could
be taken from there.